test : PackagingList.test.tsx teste à la fois l'implémentation Formik et l'implémentation RHF
