--- front/src/Apps/Forms/Components/PackagingList/FormikPackagingList.test.tsx
@@ -1,176 +1,0 @@
-import React from "react";
-import { render, fireEvent, screen, waitFor } from "@testing-library/react";
-import { Formik } from "formik";
-import { PackagingInfoInput, Packagings } from "@td/codegen-ui";
-import FormikPackagingList from "./FormikPackagingList";
-
-describe("<FormikPackagingList />", () => {
-  const defaulValues: { packagings: PackagingInfoInput[] } = {
-    packagings: [{ type: Packagings.Fut, quantity: 2 }]
-  };
-
-  const renderComponent = (initialValues?: {
-    packagings: PackagingInfoInput[];
-  }) =>
-    render(
-      <Formik
-        initialValues={initialValues ?? defaulValues}
-        onSubmit={jest.fn()}
-      >
-        <FormikPackagingList fieldName="packagings" />
-      </Formik>
-    );
-
-  it("should render correctly  with one packaging", () => {
-    renderComponent();
-    expect(screen.getByDisplayValue("Fût")).toBeInTheDocument();
-  });
-
-  it("should add a new packaging when 'Ajouter un conditionnement' button is clicked", async () => {
-    renderComponent();
-    expect(screen.getAllByLabelText("Type").length).toEqual(1);
-    // On ne peut pas supprimer le conditionnement quand il n'y en a qu'un
-    expect(screen.queryByText("Supprimer")).not.toBeInTheDocument();
-    const addButton = screen.getByText("Ajouter un conditionnement");
-    fireEvent.click(addButton);
-    await waitFor(() =>
-      expect(screen.getAllByLabelText("Type").length).toEqual(2)
-    );
-  });
-
-  it(
-    "should be possible to select Citerne or Benne as long as there" +
-      " is only one packaging type",
-    async () => {
-      renderComponent();
-      expect(screen.getByText("Citerne")).toBeInTheDocument();
-      expect(screen.getByText("Benne")).toBeInTheDocument();
-      const addButton = screen.getByText("Ajouter un conditionnement");
-      // On ajoute un deuxième conditionnement
-      fireEvent.click(addButton);
-      // Il n'est plus possible de choisir Citerne ou Benne
-      await waitFor(() => {
-        expect(screen.queryByText("Citerne")).not.toBeInTheDocument();
-        expect(screen.queryByText("Benne")).not.toBeInTheDocument();
-      });
-    }
-  );
-
-  it("Button 'Ajouter un conditionnement' should not be present when Citerne is selected", async () => {
-    renderComponent();
-    const dropdown = screen.getByLabelText("Type");
-    fireEvent.change(dropdown, { target: { value: Packagings.Citerne } });
-    await waitFor(() =>
-      expect(
-        screen.queryByText("Ajouter un conditionnement")
-      ).not.toBeInTheDocument()
-    );
-    await waitFor(() =>
-      expect(
-        screen.getByText(
-          "Un conditionnement en citerne exclut le mélange avec tout autre type de conditionnement"
-        )
-      ).toBeInTheDocument()
-    );
-  });
-
-  it("Button 'Ajouter un conditionnement' should be disabled when Benne is selected", async () => {
-    renderComponent();
-    const dropdown = screen.getByLabelText("Type");
-    fireEvent.change(dropdown, { target: { value: Packagings.Benne } });
-    await waitFor(() =>
-      expect(
-        screen.queryByText("Ajouter un conditionnement")
-      ).not.toBeInTheDocument()
-    );
-    await waitFor(() =>
-      expect(
-        screen.getByText(
-          "Un conditionnement en benne exclut le mélange avec tout autre type de conditionnement"
-        )
-      ).toBeInTheDocument()
-    );
-  });
-
-  it.each([
-    Packagings.Grv,
-    Packagings.Fut,
-    Packagings.Citerne,
-    Packagings.Autre
-  ])(
-    "should display unit in litres when packaging type is %p",
-    async packagingType => {
-      renderComponent({
-        packagings: [{ type: packagingType, quantity: 1, volume: 1 }]
-      });
-      expect(
-        screen.getByText(`Volume en litres (optionnel)`)
-      ).toBeInTheDocument();
-      expect(screen.getByText(`Soit 0.001 m3`)).toBeInTheDocument();
-    }
-  );
-
-  it("should convert volume from litres to m3  when packaging type is benne", async () => {
-    renderComponent({
-      packagings: [
-        {
-          type: Packagings.Benne,
-          quantity: 1,
-          // le volume reçu est en litres
-          volume: 1000
-        }
-      ]
-    });
-    // L'affichage doit se faire en m3
-    const volumeInputInM3 = screen.getByLabelText("Volume en m3 (optionnel)");
-    expect(volumeInputInM3).toBeInTheDocument();
-    expect(volumeInputInM3).toHaveValue(1);
-    expect(screen.getByText(`Soit 1000 litres`)).toBeInTheDocument();
-  });
-
-  it("should remove a packaging when 'Supprimer' button is clicked", async () => {
-    renderComponent({
-      packagings: [
-        { type: Packagings.Fut, quantity: 2 },
-        { type: Packagings.Grv, quantity: 1 }
-      ]
-    });
-    expect(
-      screen.getByDisplayValue("Grand Récipient Vrac (GRV)")
-    ).toBeInTheDocument();
-    const deleteButtons = screen.getAllByText("Supprimer");
-    expect(deleteButtons).toHaveLength(2);
-    fireEvent.click(deleteButtons[1]);
-    await waitFor(() => {
-      expect(
-        screen.queryByDisplayValue("Grand Récipient Vrac (GRV)")
-      ).not.toBeInTheDocument();
-    });
-  });
-
-  it("should display input 'Nom du type de conditionnement' when Type 'Autre' is selected", async () => {
-    renderComponent({
-      packagings: [{ type: Packagings.Autre, quantity: 1 }]
-    });
-    expect(
-      screen.getByLabelText("Nom du type de conditionnement")
-    ).toBeInTheDocument();
-  });
-
-  it("should display identification numbers", async () => {
-    renderComponent({
-      packagings: [
-        {
-          type: Packagings.Fut,
-          quantity: 2,
-          identificationNumbers: ["identifiant1", "identifiant2"]
-        }
-      ]
-    });
-    expect(screen.getByText("identifiant1")).toBeInTheDocument();
-    expect(screen.getByText("identifiant2")).toBeInTheDocument();
-    expect(
-      screen.getByText("Vous avez saisi 2 numéros pour 2 contenants")
-    ).toBeInTheDocument();
-  });
-});
